Change getstorybook command to storybook init
